+++ kmeans_functions.py
@@ -1,5 +1,4 @@
 import numpy as np
-import scipy.spatial.distance as ssdist
 import initialization as init
 
 
@@ -14,7 +13,7 @@ def initialization(n_clusters, initialize, x):
 
 def labeling(centroids, x):
     # this method creates labels for every position in training data based on current centroids
-    distance = fix_the_distance(x, centroids, 'correlation')
+    distance = init.fix_the_distance(x, centroids, 'correlation')
     labels = np.argmin(distance, axis=1)
     return labels
 
@@ -39,16 +38,6 @@ def _centroid_by_means(n_clusters, centroids, labels, x):
 
 def inertia(centroids, x):
     datasize = x.shape[0]
-    distance = fix_the_distance(x, centroids, 'correlation')
+    distance = init.fix_the_distance(x, centroids, 'correlation')
     inertia = np.sum([(distance[_]) ** 2 for _ in range(datasize)])
     return inertia
-
-
-def fix_the_distance(x, centroids, dist_metric):
-    distance = ssdist.cdist(x, centroids, metric=dist_metric)  # here is sth to change
-    distance[np.isnan(distance)] = 0
-    for i in range(x.shape[0]):
-        for j in range(centroids.shape[0]):
-            if (x[i] == centroids[j]).all():
-                distance[i, j] = 0
-    return distance
